requirements.txt
# python 3.8.20
# torch==2.4.1+cu124
# torchaudio==2.4.1+cu124
# torchvision==0.19.1+cu124
darts==0.25.0
matplotlib>=3.6.2
numpy>=1.22.3
pandas>=1.5.1
scikit_learn>=1.1.3
scipy>=1.8.0
statsmodels>=0.14.0
ray>=2.6.3
tqdm>=4.64.0
dash>=2.9.3
dash-bootstrap-components>=1.5.0
reformer-pytorch==1.4.4
lightgbm>=4.1.0
shap>=0.41.0
# cvxopt
# pathos